requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.3.4
redis==5.0.6
rq==1.16.2
pillow==10.3.0
opencv-python-headless==4.10.0.82
realesrgan==0.3.0
# torch CPU version will be handled in Dockerfile or pip specifically to avoid huge GPU bloat if possible, 
# but for simplicity we list standard torch here, pip often grabs cuda.
# Better to let Dockerfile handle complex pytorch install if needed.
torch
torchvision
basicsr>=1.4.2
streamlit==1.36.0
requests==2.32.3
aiofiles==24.1.0
numpy
